+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Version History
 
-- 0.1.0: Added [pyawsopstoolkit.exceptions](pyawsopstoolkit/exceptions) package. (latest)
+- 0.1.1: Added [pyawsopstoolkit.validators](pyawsopstoolkit/validators) package. (latest)
+- 0.1.0: Added [pyawsopstoolkit.exceptions](pyawsopstoolkit/exceptions) package.